Create subfolders for main components and renaming
Reorganize folder structure for clarity. Rename components for enhanced
readability.

- Rename ExpressionInputComponent -> ExpressionEditComponent
- Rename VariablesTable->VectorSelectorTable
- Update VectorCalculator.css naming accordingly
- Create subfolders for ExpressionEditComponent and ExpressionsTableComponent
- Place all all component-specific content within folders
