+++ react/src/lib/components/VectorCalculator/components/ExpressionEditComponent/components/ExpressionNameTextField.tsx
@@ -3,15 +3,15 @@ import { error_filled, warning_filled, thumbs_up } from "@equinor/eds-icons";
 import { TextField, Icon } from "@equinor/eds-core-react";
 import { TreeDataNode } from "@webviz/core-components";
 
-import { StoreActions, useStore } from "./ExpressionsStore";
+import { StoreActions, useStore } from "../../ExpressionsStore";
 import {
     isNameOccupiedByVectors,
     doesNameExistInExpressionList,
     isValidExpressionNameString,
     expressionNameValidationMessage,
-} from "../utils/VectorCalculatorHelperFunctions";
+} from "../../../utils/VectorCalculatorHelperFunctions";
 
-import "!style-loader!css-loader!../VectorCalculator.css";
+import "!style-loader!css-loader!../../../VectorCalculator.css";
 
 type ExpressionNameTextFieldVariantType =
     | "success"
