+++ react/src/lib/components/VectorCalculator/components/ExpressionEditComponent/index.ts
@@ -1,0 +1,1 @@
+export { ExpressionEditComponent } from "./ExpressionEditComponent";
