+++ react/src/lib/components/VectorCalculator/components/ExpressionsTableComponent/components/EnhancedTableHead.tsx
