+++ react/src/lib/components/VectorCalculator/components/ExpressionsTableComponent/index.ts
@@ -1,0 +1,1 @@
+export { ExpressionsTableComponent } from "./ExpressionsTableComponent";
