--- react/src/lib/components/VectorCalculator/components/MaxLengthTextField/index.tsx
@@ -1,1 +1,0 @@
-export { MaxLengthTextField } from "./MaxLengthTextField";
