+++ react/src/lib/components/VectorCalculator/utils/ConfirmDialog.tsx
